scream: Use GstAtomicQueue for stream packet queue
The stream packet queue can be accessed by multiple threads, so we need
to make sure that checking its length and then popping is an atomic
operation. Since this is hard to do with GstDataQueue (which makes
popping and peeking synchronous), we use a GstAtomicQueue instead, which
is also simpler.