gstsctpdec: Use activation callback for pads to stop task
This patch fixes a deaclock when the transition from PAUSE to READY takes place
and pads are implicitly deactivated during the state_change function. When pads
are deactived during the state transtion, the decoder task is locked in the
gst_data_queue_pop so the transition to READY can never success blocking the
thread when it tries to acquired the mutex of that pad.
This patch uses the activation function to manage the life cicle of the source
pad task, when the change of state to READY takes place, the pad is implicitly
deactivated using this callback, here we make sure that task is unlocked and
stoped without causing any deadlocks.